[#67] refactor: Layout 컴포넌트를 layout 디렉토리로 이동
