+++ src/pages/HomePage/index.tsx
@@ -1,6 +1,6 @@
 import React, { FC } from 'react';
 
-import Layout from '@/components/common/Layout';
+import Layout from '@/components/layout/Layout';
 import InduceSignup from '@/components/intro/InduceSignup';
 import IntroMain from '@/components/intro/IntroMain';
 import IntroWithClientUI from '@/components/intro/IntroWithClientUI';
